☕ Moved validation class to core validator
